bay vong tron

+++ src/model/Model.java
@@ -4,8 +4,8 @@ package model;
  * Created by Dell on 17/12/2016.
  */
 public class Model {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private int width;
     private int height;
 
@@ -18,7 +18,7 @@ public class Model {
         return height;
     }
 
-    public Model(int x, int y, int width, int height) {
+    public Model(double x, double y, int width, int height) {
 
         this.x = x;
         this.y = y;
@@ -26,7 +26,7 @@ public class Model {
         this.height = height;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
@@ -34,7 +34,7 @@ public class Model {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
